updated init file.
As such the init file should be generated from the org file. This
setup has to be revised such that the init.el never has to be commited
and remains stable. Since the org file contains the real updates to
the config and is the file that I want available as a literate config,
the init file's commits make the commit structure messy.

Earlier, I used to have an init.el with the bare minimum and so it
hardly changed. This init file would load the basic libraries needed
for calling the org file. This would then mean having atleast 2 elisp
scripts for calling emacs.