Major refactor: separate init.el and dotemacs.el
This creates a source of truth org file which is tangled to init.el
and a dotemacs.el.

init.el contains a bootstrap of straight and use-package and minimal
functions to enable loading the dotemacs.el file. The next plan is to
tangle the extended config, if it is not available (i.e catering to a
new emacs setup).

The dotemacs.el file contains all the main configuration. This file
will not be included in git commits as it is always in a state of flux
and it makes no sense to commit the same portions of code in the org
file as well as the el file.

The idea behind the above is that an init.el will exist to load Emacs
quickly in any new setup, and it will remain stable with minimal
changes, if at all. This enables a clear strategy of making commits
only of the source-of-truth org file. Essentially, it will enable a
meaningful history of changes to the extent that I can deduce at the
moment.

Probably worth noting that the dotemacs.el will also load a secret and
encrypted config file if it exists. As of now, the secret file is not
included in the public commits of the git repo.