Adds org-journal into the main configuration.
Since I have now changed the file format to be a little more organised,
it will not be possible to include the previous entries in the calendar
without modifications of the file name and so on. In any case, it is
more important to enable a search through the entire journal folder
alone. This can be done using helm-do-ag