Removes inactive date for active link entries
The inactive date can be called as a filtered view in the
org-agenda. Therefore, when there is an active date in addition - it
would show up twice in the agenda. The created date property is to
understand when the task was created. Other than refile tasks, there is
no need to enter both an active and an inactive date for a
headline. Instead, the date can be passed to the created property as a string.