Adds capture template for file links
The idea is that files from org-brain can be captured as tasks and then
clocked. An attempt was made to use the org-web-tools package for
capturing external webpage links to read. However, I was unable to
correctly decipher how to get the function to work. This has been
commited for now.