org journal - no carry over for items
carrying over items is useful, however, in the past this has often
resulted in item duplication. This is because of mulitple reasons - I am
often jumping across journal files and multiple journal files are
open. Additionally, if one is often working at midnight when the day is
changing, and the file is not saved and then a new journal command is
called the next day - it results sometimes in duplication of content
being carried over.