Enabling the agignore option for helm-ag
Having ag search through ipynb files and some specific useless
directories in a project is annoying. However, excluding anything from
the search has to be done in a cautious manner. This option enables the
agignore file. It can be disabled quickly as well with a setq to
nil. There are also other options to be explored with respect to larger
coding projects and a flexible method of ignoring files and
directories. I think this approach should be simple and standard so that
I can easily remember the same whenever I am searching through a
project.