Helm and counsel changes - re-init helm
I had been using counsel for most things and helm for a few
functions. With the GC optimisation borrowed from doom, helm seems
responsive enough. I like being able to quickly look at additional
options with a simple C-j from anywhere. The lists are also more
comprehensive and more suitable to my wandering mind. Nevertheless,
counsel's simple interface is certainly useful in many ways. The goal is
to find an optimal sweet spot between counsel and helm for my daily
workflow.

In particular, at the moment, helm has a large number of options
in-built, that are very useful. For example, it is easy to view the
files in another project while the current project is open and mark and
delete a few files. THere are other options of ediff and merge and etc
readily available that I would like to be able to use. The issue at the
moment is that the default key-bindings for helm are not activated. This
will be rectified down the line.