winum scope frame local
If each frame is restricted to 2-3 windows then it is easy to switch
between them using C-1, C-2, C-3 and even C-4. More than these, it
becomes too taxing to switch between the buffers to be productive. For
such situations, I would prefer using eyebrowse, but still trying to
limit the number of buffers open at any one time should not exceed the
ability to switch between them and maintain coherence of thought.