Disable habits during agenda startup
Habits can be displayed using the k shortcut when the agenda is
called. This is currently done as an attempt to declutter the agenda and
enable actually using the agenda as one would imagine. So far it seems
that a decluttered agenda is the best way to get this done. While it is
great to be able to record everything with a time stamped basis, there
should be a way to filter out items quickly in the agenda and also be
able to quickly jump between the agenda and tasks. The follow mode in
the agenda can help with this. Another reason to disable the habits
display is to test with the on going troubleshooting of org agenda
responsiveness.