Testing longer hydra shortcuts for functions
though it is easiest to have 2 keys for hydras, or a series of single
keys, sometimes it is necessary to chain together or cluster a bunch of
commands together. In thise case, the which key provides, some
information about the options available, and is in a way a set of single
key presses anyway.