Adds perspective-el and persp-projectile
Learning notes from perspective are captured in org-brain and will be
published later. As of now it is important to understand that
perspective deals with lists of buffers and configurations. For this the
buffers or files have to exist. Therefore processes like calling mu4e or
the git status cannot be done using perspectives. This can be defined in
functions that include window configurations though.