Sets a default perspective file
This is much more convenient because the perspective files can then be
loaded quickly, without navigating to the location each time. THis is
particularly cumbersome in perspective-el when dealing with multiple
frames running on an Emacs daemon. The better solution may be to use the
persp-mode package that allows multiple frames.