House keeping and several hydra updates
Adds a hydra-jump meant to serve as a one stop location to jump around
in code files, and org files. The journal can be directly called with
H-j which is more convenient.

Adds org-journal hydras to navigate the org journal buffers both between
daily entries and also between headlines in each buffer. THis has to be
streamlined a bit more and the key configuration has to be improved.

Adds the function of inserting captured messages in mu4e in an email in
a hydra. This makes it very easy to capture any set of random
messages (using c in the mu4e header view) and attach them all to an
email.

Optimised the perspective hydra. I am trying to use '[' ']' for moving
back and forward in packages like perspective-el, winner and eyebrowse.