Scimax integration update - adds the forked repo
Though this is still messy - this integration has worked so far as
expected for me. This commit adds the sr.ht repo which contains a
smaller subset of the scimax repo as it was several months ago. Each
feature is added as I find that I need something that I was using in
Scimax. I am not yet certain this is the best way to do things.