ADds org id to seveal headlines, org-roam
several changes to the file and folder structure has been created with
the idea of integrating org-roam into the workflow. The paths have been
changed throuhgout the Emacs configuration appropriately. The idea is
that most things will remain private and as they are, but are enclosed
within a public folder. org-roam enables inheritance of tags based on
the last directory.