Modifies gists and webpaste hydra to include git gutter
This hydra is somewhat useful in quickly staging hunks and moving
between chunks. The movement, marking and etc are not entirely smooth.
A hook should be added to save the document prior to calling the
hydra. I find it easier to do this rather looking for the chunk among
the many diffs.